# src/main/react/distrupify/openapi.yaml
...
openapi: 3.0.3
info:
  title: distrupify API
  version: 1.0.0-SNAPSHOT
paths:
...
  /api/v1/suppliers:
    get:
      tags:
      - Supplier Resource
      parameters:
      - name: page
        in: query
        schema:
          format: int32
          type: integer
      - name: page_size
        in: query
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuppliersResponse'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
    post:
      tags:
      - Supplier Resource
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierCreateRequest'
      responses:
        "200":
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/v1/suppliers/{id}:
    put:
      tags:
      - Supplier Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierEditRequest'
      responses:
        "200":
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
    delete:
      tags:
      - Supplier Resource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
components:
  schemas:
...
    SupplierCreateRequest:
      required:
      - name
      - address
      - contactNumber
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        contactNumber:
          type: string
    SupplierDTO:
      required:
      - id
      - name
      - address
      - contactNumber
      type: object
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        address:
          type: string
        contactNumber:
          type: string
    SupplierEditRequest:
      required:
      - name
      - address
      - contactNumber
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        contactNumber:
          type: string
    SuppliersResponse:
      required:
      - suppliers
      - pageCount
      type: object
      properties:
        suppliers:
          type: array
          items:
            $ref: '#/components/schemas/SupplierDTO'
        pageCount:
          format: int32
          type: integer
...
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT